feature: add validation bussiness in use cases

+++ src/application/use-cases/recipe-step/delete-recipe-step.ts
@@ -1,8 +1,8 @@
-import { UniqueEntityID } from "../../../core/domain/value-objects/unique-entity-id";
 import { Either, failure, success } from "../../../core/either";
 import { NotFoundError } from "../../errors/resource-not-found-error";
 import { RecipeStepRepository } from "../../repositories/recipe-step-repository";
 import { NotAllowedError } from "../../errors/not-allowed-error";
+import { RecipeRepository } from "../../repositories/recipe-repository";
 
 interface DeleteRecipeStepUseCaseRequest {
   id: string;
@@ -12,7 +12,10 @@ interface DeleteRecipeStepUseCaseRequest {
 type DeleteRecipeStepUseCaseResponse = Either<NotFoundError | NotAllowedError, null>;
 
 export class DeleteRecipeStepUseCase {
-  constructor(private recipeStepRepository: RecipeStepRepository) {}
+  constructor(
+    private recipeStepRepository: RecipeStepRepository,
+    private recipeRepository: RecipeRepository,
+  ) {}
   async execute({
     id,
     deletedBy,
@@ -26,6 +29,15 @@ export class DeleteRecipeStepUseCase {
       return failure(new NotAllowedError("user"));
     }
 
+    const recipe = await this.recipeRepository.findById(recipeStep.recipeId.toString());
+
+    if (!recipe) {
+      return failure(new NotFoundError("recipe"));
+    }
+
+    if (recipe?.status.toString() !== "ACTIVE") {
+      return failure(new NotAllowedError("recipe"));
+    }
     await this.recipeStepRepository.delete(recipeStep);
 
     return success(null);
